Moved remaining quota calculation from cell quota dm to cell quota service
