+++ src/casp/services/api/data_manager/cell_quota.py
@@ -16,31 +16,6 @@ class CellQuotaDataManager(BaseDataManager):
     with open(f"{SQL_QUERY_DIR}/get_cells_processed_this_week_for_user.sql.sa", "r") as f:
         SQL_GET_CELLS_PROCESSED_THIS_WEEK = f.read()
 
-    def get_remaining_quota_for_user(self, user: models.User) -> int:
-        """
-        Get remaining cell processing quota for a specific user (i.e. their quota minus cells processed this week
-        or their remaining lifetime quota if they have one)
-
-        :param user: User object to check quota for
-
-        :return: Remaining cell processing quota
-        """
-        # Return the smaller of the remaining weekly and lifetime quotas, or 0 if it's negative
-        # (which might happen if the user exceeds their quota, either because we allow them to, or
-        # because of any weirdness in the simultaneous processing of cells)
-        weekly_remaining_quota = user.cell_quota - self.get_cells_processed_this_week_for_user(user=user)
-
-        lifetime_remaining_quota = (
-            user.lifetime_cell_quota - user.total_cells_processed
-            if user.lifetime_cell_quota is not None
-            else float("inf")
-        )
-
-        remaining_quota = min(weekly_remaining_quota, lifetime_remaining_quota)
-
-        # Ensure the remaining quota is not negative
-        return max(remaining_quota, 0)
-
     def get_cells_processed_this_week_for_user(self, user: models.User) -> int:
         """
         Get number of cells processed by a specific user this week
